Improve import of database resource and add guide to migrate from old to new provider
